Consolidate download scripts into utils.py and use requests with tqdm instead of wget package
